requirements.txt
imgui==2.0.0
matplotlib==3.8.2
moderngl==5.10.0
moderngl_window==2.4.6
numpy==1.24.1
opencv_contrib_python==4.9.0.80
opencv_python==4.8.1.78
Pillow==10.4.0
pyrr==0.10.3
